+++ test/starkbank/dict_key_test.clj
@@ -13,6 +13,6 @@
 (deftest query-dict-key
 	(testing "query dict-key"
 		(user/set-test-user)
-		(def dict-keys (take 200 (dict-key/query {:limit 1, :status ["registered" "created"] :type "evp"})))
+		(def dict-keys (take 200 (dict-key/query {:limit 1 :status ["registered" "created"] :type "evp"})))
 		(is (= 1 (count dict-keys)))
 		(is (not (nil? (:id (first dict-keys)))))))